% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{india_china_males_1990}
\alias{india_china_males_1990}
\title{Age and Cause Decomposition of Difference in Life Expectancies at Birth, India and China, males, 1990}
\format{
A data frame with 7 rows and 10 columns:
\describe{
\item{Age}{ordered age groups as factor type}
\item{India_nmx}{all-cause mortality rate between ages x and x + n for males in India, 1990}
\item{India_CD}{proportion of deaths from communicable diseases between ages x and x + n for males in India, 1990}
\item{India_NCD}{proportion of deaths from non-communicable diseases between ages x and x + n for males in India, 1990}
\item{India_Injuries}{proportion of deaths from injuries between ages x and x + n for males in India, 1990}
\item{China_nmx}{all-cause mortality rate between ages x and x + n for males in China, 1990}
\item{China_CD}{proportion of deaths from communicable diseases between ages x and x + n for males in China, 1990}
\item{China_NCD}{proportion of deaths from non-communicable diseases between ages x and x + n for males in China, 1990}
\item{China_Injuries}{proportion of deaths from injuries between ages x and x + n for males in China, 1990}
\item{nDx}{contribution of all-cause mortality differences in groups 1 and 2 in age groups x to x + n. Computationally the same as the total effect column computed from \code{decomp_LE()}.}

where \emph{n} = length of the age interval
}
}
\source{
Murray, C. J. and A. D. Lopez, 1996. \emph{The Global Burden of Disease: A Comprehensive Assessment of Mortality and Disability from Diseases, Injuries, and Risk Factors in 1990 and Projected to 2020}. Boston, Harvard University, School of Public Health.
}
\usage{
india_china_males_1990
}
\description{
A subset of data from The Global Burden of Disease Study 1996.
}
\keyword{datasets}
