% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{india_china_males_1995}
\alias{india_china_males_1995}
\title{Age and Cause Decomposition of Difference in Life Expectancies at Birth, India and China, males, 1990}
\format{
A data frame with 7 rows and 13 columns:
\describe{
\item{Age}{ordered age groups as factor type}
\item{nm1x}{age specific death rate in the age group x; x + n in the initial time period ‘1935’}
\item{l1x}{number of persons alive at exact age x, in the initial time period ‘1935’ expressed as decimal form and as a proportion of the starting the age group}
\item{e1x}{expectation of life at exact age x, in the initial time period ‘1935’}
\item{nm2x}{age specific death rate in the age group x; x + n in the latter time period ‘1995’}
\item{l2x}{number of persons alive at exact age x, in the latter time period ‘1995’ expressed as decimal form and as a proportion of the starting the age group}
\item{e2x}{expectation of life at exact age x, in the latter time period ‘1995’}

where \emph{n} = length of the age interval
}
}
\source{
Murray, C. J. and A. D. Lopez, 1996. \emph{The Global Burden of Disease: A Comprehensive Assessment of Mortality and Disability from Diseases, Injuries, and Risk Factors in 1990 and Projected to 2020}. Boston, Harvard University, School of Public Health.
}
\usage{
india_china_males_1995
}
\description{
A subset of data from The Global Burden of Disease Study 1996.
}
\keyword{datasets}
