% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decomposition.R
\name{decomposition}
\alias{decomposition}
\title{Title decomposition}
\usage{
decomposition(df, method = "arriaga3", age_col, e1, e2, l1, l2, append = TRUE)
}
\arguments{
\item{df}{An outputted life table with}

\item{method}{Methods to use for life expectancy decomposition. Defaults to 'arriaga3'. Current methods available are: 'arriaga3'.}

\item{age_col}{Column providing ordered age bands with the final age group being an open-ended interval suffxied with '+', e.g. '90+'.. Of factor type.}

\item{e1}{Column name for expectation of life at age group x, in the 1st group of comparison.}

\item{e2}{Column name for expectation of life at age group x, in the second group of comparison.}

\item{l1}{Column name for The number of persons alive at age group x, in the 1st group of comparison.}

\item{l2}{Column name for The number of persons alive at age group x, in the 1st group of comparison.}

\item{append}{Whether to append the decomposition columns to the original data frame.}
}
\value{
A data frame with life expectancy decomposition values
}
\description{
description Function for performing life expectancy decomposition.
}
\examples{
decomposition(us_females, age_col = "Age", e1 = "e1x", e2 = "e2x", l1 = "l1x", l2 = "l2x")
}
